[RefData][Geonames] New Geonames data file format for the alternate names.
The section of alternate names now lists them with a pipe (|) separator,
instead of a hat (^) one. Hence, the number of fields of the Geonames-based
data files is now constant. The Shell/AWK scripts have been amended accordingly.

For now, the (output) format of the ORI-maintained list of POR remains the same,
so as to easily compare the data files and thus to secure the migration.
The new format will be activated in the next version."